feat+: major UI updates, multilingual, bugs fixed, new calorie tracker
**Description:**
- Added fast scroll with alphabetical popup and improved FAB spacing
- Implemented Spanish/French localization
- Highlight color fixed; popular sort corrected; muscle sort refined
- Added tutorial dialog, half‑weight support, exercise detail expansion
- Created calories tracker, notes UI, body status, user avatar
- Added import/search/filter, copy workout, heatmap view
- Fixed multiple bugs (kcal save, import duplication, index resets, etc.)
- DB migrations 14→17; many UI/adapter enhancements

+++ app/src/main/java/com/example/gymevo/data/seed/WorkoutSeed.java
@@ -324,7 +324,7 @@ public final class WorkoutSeed {
                 exercise.getImageB(),
                 series,
                 repetitions,
-                weight,
+                (float) weight,
                 time,
                 heartRates,
                 null,
